gcp-emissions-project/requirements.txt
pandas
numpy
google-cloud-storage
google-cloud-bigquery
pyarrow
